requirements.txt
# Build dependencies
setuptools>=65.0.0
wheel>=0.37.0

# Core ML and LLM libraries
langchain==0.1.0
langchain-community==0.0.13
transformers==4.36.2
torch>=2.2.0
sentence-transformers==2.2.2

# Vector Database
pinecone>=3.0.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4

# Database Connectivity
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
mysql-connector-python>=8.0.0

# Data Visualization
matplotlib==3.8.2
plotly==5.17.0
seaborn==0.13.0

# Web Framework
streamlit>=1.28.0
streamlit-option-menu>=0.3.0
streamlit-aggrid>=0.3.0

# Text Processing
nltk==3.8.1
spacy==3.7.2

# Utility Libraries
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.4
scikit-learn==1.3.2

# Re-ranking
rank-bm25==0.2.2

# UI Components
streamlit-ace==0.1.1
streamlit-chat==0.1.1

# Additional utilities
Pillow==10.1.0
python-magic==0.4.27